+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'dart:io';
 import 'dart:ffi';
 import 'dart:math';
 import 'dart:typed_data';
@@ -23,8 +24,8 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final _textureRgbaRendererPlugin = TextureRgbaRenderer();
   int textureId = -1;
-  int height = 500;
-  int width = 500;
+  int height = 768;
+  int width = 1377;
   int cnt = 0;
   var key = 0;
   int texturePtr = 0;
@@ -33,6 +34,7 @@ class _MyAppState extends State<MyApp> {
   Timer? _timer;
   int time = 0;
   int method = 0;
+  final strideAlign = Platform.isMacOS ? 64 : 1;
 
   @override
   void initState() {
@@ -58,30 +60,32 @@ class _MyAppState extends State<MyApp> {
   void start(int methodId) {
     debugPrint("start mockPic");
     method = methodId;
+    final rowBytes = (width * 4 + strideAlign - 1) & (~(strideAlign - 1));
+    final picDataLength = rowBytes * height;
+    debugPrint('REMOVE ME =============================== rowBytes $rowBytes');
     _timer?.cancel();
     // 60 fps
     _timer =
         Timer.periodic(const Duration(milliseconds: 1000 ~/ 60), (timer) async {
       if (methodId == 0) {
         // Method.1: with MethodChannel
-        data = mockPicture(width, height);
+        data = mockPicture(width, height, rowBytes, picDataLength);
         final t1 = DateTime.now().microsecondsSinceEpoch;
-        final res =
-            await _textureRgbaRendererPlugin.onRgba(key, data!, height, width);
+        final res = await _textureRgbaRendererPlugin.onRgba(
+            key, data!, height, width, strideAlign);
         final t2 = DateTime.now().microsecondsSinceEpoch;
         setState(() {
           time = t2 - t1;
         });
-
         if (!res) {
           debugPrint("WARN: render failed");
         }
       } else {
-        final dataPtr = mockPicturePtr(width, height);
+        final dataPtr = mockPicturePtr(width, height, rowBytes, picDataLength);
         // Method.2: with native ffi
         final t1 = DateTime.now().microsecondsSinceEpoch;
         Native.instance.onRgba(Pointer.fromAddress(texturePtr).cast<Void>(),
-            dataPtr, width, height);
+            dataPtr, picDataLength, width, height, strideAlign);
         final t2 = DateTime.now().microsecondsSinceEpoch;
         setState(() {
           time = t2 - t1;
@@ -100,16 +104,55 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
   }
 
-  Uint8List mockPicture(int width, int height) {
-    final pic = List.generate(width * height * 4, (index) {
-      return random.nextInt(255);
+  Uint8List mockPicture(int width, int height, int rowBytes, int length) {
+    final pic = List.generate(length, (index) {
+      final r = index / rowBytes;
+      final c = (index % rowBytes) / 4;
+      final p = index & 0x03;
+      if (c > 20 && c < 30) {
+        if (r > 20 && r < 25) {
+          if (p == 0 || p == 3) {
+            return 255;
+          } else {
+            return 0;
+          }
+        }
+        if (r > 40 && r < 45) {
+          if (p == 1 || p == 3) {
+            return 255;
+          } else {
+            return 0;
+          }
+        }
+        if (r > 60 && r < 65) {
+          if (p == 2 || p == 3) {
+            return 255;
+          } else {
+            return 0;
+          }
+        }
+      }
+      return 255;
     });
     return Uint8List.fromList(pic);
   }
 
-  Pointer<Uint8> mockPicturePtr(int width, int height) {
-    final pic = List.generate(width * height * 4, (index) {
-      return random.nextInt(255);
+  Pointer<Uint8> mockPicturePtr(
+      int width, int height, int rowBytes, int length) {
+    final pic = List.generate(length, (index) {
+      final r = index / rowBytes;
+      final c = (index % rowBytes) / 4;
+      final p = index & 0x03;
+      final edgeH = (c >= 0 && c < 10) || ((c >= width - 10) && c < width);
+      final edgeW = (r >= 0 && r < 10) || ((r >= height - 10) && r < height);
+      if (edgeH || edgeW) {
+        if (p == 0 || p == 3) {
+          return 255;
+        } else {
+          return 0;
+        }
+      }
+      return 255;
     });
     final picAddr = malloc.allocate(pic.length).cast<Uint8>();
     final list = picAddr.asTypedList(pic.length);
@@ -121,41 +164,40 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        body: Center(
-          child: SingleChildScrollView(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                textureId == -1
-                    ? const Offstage()
-                    : Container(
-                        alignment: Alignment.center,
-                        decoration: const BoxDecoration(color: Colors.blue),
-                        // decoration: const BoxDecoration(color: Colors.black),
-                        width: width.toDouble(),
-                        height: height.toDouble(),
-                        child: Texture(textureId: textureId)),
-                Text(
-                    "texture id: $textureId, texture memory address: ${texturePtr.toRadixString(16)}"),
-                TextButton.icon(
-                  label: const Text("play with texture (method channel API)"),
-                  icon: const Icon(Icons.play_arrow),
-                  onPressed: () => start(0),
-                ),
-                TextButton.icon(
-                  label: const Text("play with texture (native API, faster)"),
-                  icon: const Icon(Icons.play_arrow),
-                  onPressed: () => start(1),
-                ),
-                Text(
-                    "Current mode: ${method == 0 ? 'Method Channel API' : 'Native API'}"),
-                time != 0
-                    ? Text("FPS: ${1000000 ~/ time} fps")
-                    : const Offstage()
-              ],
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Expanded(
+              child: textureId == -1
+                  ? const Offstage()
+                  : Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: Container(
+                          alignment: Alignment.center,
+                          decoration: const BoxDecoration(color: Colors.blue),
+                          // decoration: const BoxDecoration(color: Colors.black),
+                          // width: width.toDouble(),
+                          // height: height.toDouble(),
+                          child: Texture(textureId: textureId)),
+                    ),
             ),
-          ),
+            Text(
+                "texture id: $textureId, texture memory address: ${texturePtr.toRadixString(16)}"),
+            TextButton.icon(
+              label: const Text("play with texture (method channel API)"),
+              icon: const Icon(Icons.play_arrow),
+              onPressed: () => start(0),
+            ),
+            TextButton.icon(
+              label: const Text("play with texture (native API, faster)"),
+              icon: const Icon(Icons.play_arrow),
+              onPressed: () => start(1),
+            ),
+            Text(
+                "Current mode: ${method == 0 ? 'Method Channel API' : 'Native API'}"),
+            time != 0 ? Text("FPS: ${1000000 ~/ time} fps") : const Offstage()
+          ],
         ),
       ),
     );