+++ linux/texture_rgba_renderer_plugin.cc
@@ -12,11 +12,13 @@
 #include <cstring>
 #include <unordered_map>
 
+std::unordered_map<int64_t, TextureRgba *> g_renderer_map;
+
+void FlutterRgbaRendererPluginOnRgba(void *texture_rgba_ptr, const uint8_t *buffer, int len, int width, int height);
 struct _TextureRgbaRendererPlugin
 {
   GObject parent_instance;
   FlTextureRegistrar *texture_registrar;
-  std::unordered_map<int64_t, TextureRgba *> *g_renderer_map;
 };
 
 G_DEFINE_TYPE(TextureRgbaRendererPlugin, texture_rgba_renderer_plugin, g_object_get_type())
@@ -30,29 +32,20 @@ static void texture_rgba_renderer_plugin_handle_method_call(
 
   const gchar *method = fl_method_call_get_name(method_call);
 
-  if (strcmp(method, "getPlatformVersion") == 0)
-  {
-    struct utsname uname_data = {};
-    uname(&uname_data);
-    g_autofree gchar *version = g_strdup_printf("Linux %s", uname_data.version);
-    g_autoptr(FlValue) result = fl_value_new_string(version);
-    response = FL_METHOD_RESPONSE(fl_method_success_response_new(result));
-  }
-  else if (strcmp(method, "createTexture") == 0)
+  if (strcmp(method, "createTexture") == 0)
   {
     auto args = fl_method_call_get_args(method_call);
     int64_t key = fl_value_get_int(fl_value_lookup_string(args, "key"));
-    if (self->g_renderer_map->find(key) != self->g_renderer_map->end())
+    if (g_renderer_map.find(key) != g_renderer_map.end())
     {
       response = FL_METHOD_RESPONSE(fl_method_success_response_new(fl_value_new_int(-1)));
     }
     else
     {
-      g_autoptr(TextureRgba) texture_rgba = texture_rgba_new(self->texture_registrar);
+      TextureRgba* texture_rgba = texture_rgba_new(self->texture_registrar);
       auto texture_id = reinterpret_cast<int64_t>(FL_TEXTURE(texture_rgba));
-      FL_TEXTURE_GL_GET_CLASS(texture_rgba)->populate =
-          texture_rgba_populate;
-      self->g_renderer_map->insert(std::make_pair(key, texture_rgba));
+      texture_rgba->texture_id = texture_id;
+      g_renderer_map.insert(std::make_pair(key, texture_rgba));
       // Register to the registrar.
       fl_texture_registrar_register_texture(self->texture_registrar, FL_TEXTURE(texture_rgba));
       response = FL_METHOD_RESPONSE(fl_method_success_response_new(fl_value_new_int(texture_id)));
@@ -62,10 +55,11 @@ static void texture_rgba_renderer_plugin_handle_method_call(
   {
     auto args = fl_method_call_get_args(method_call);
     int64_t key = fl_value_get_int(fl_value_lookup_string(args, "key"));
-    if (self->g_renderer_map->find(key) != self->g_renderer_map->end())
+    if (g_renderer_map.find(key) != g_renderer_map.end())
     {
-      fl_texture_registrar_unregister_texture(self->texture_registrar, FL_TEXTURE((*self->g_renderer_map)[key]));
-      self->g_renderer_map->erase(key);
+      fl_texture_registrar_unregister_texture(self->texture_registrar, FL_TEXTURE(g_renderer_map[key]));
+      texture_rgba_terminate(g_renderer_map[key]);
+      g_renderer_map.erase(key);
     }
     response = FL_METHOD_RESPONSE(fl_method_success_response_new(fl_value_new_bool(true)));
   }
@@ -75,20 +69,11 @@ static void texture_rgba_renderer_plugin_handle_method_call(
     auto key = fl_value_get_int(fl_value_lookup_string(args, "key"));
     auto data = fl_value_get_uint8_list(fl_value_lookup_string(args, "data"));
     auto data_length = fl_value_get_length(fl_value_lookup_string(args, "data"));
-    auto buffer = new uint8_t[data_length];
-    memcpy(buffer, data, data_length);
     auto width = fl_value_get_int(fl_value_lookup_string(args, "width"));
     auto height = fl_value_get_int(fl_value_lookup_string(args, "height"));
-    auto texture_rgba = (*self->g_renderer_map)[key];
-    auto texture_rgba_private = (TextureRgbaPrivate *)texture_rgba_get_instance_private((*self->g_renderer_map)[key]);
-    g_mutex_lock(&texture_rgba_private->mutex);
-    if (texture_rgba_private->buffer != nullptr) {
-      delete[] texture_rgba_private->buffer;
-    } 
-    texture_rgba_private->buffer = buffer;
-    texture_rgba_private->video_height = height;
-    texture_rgba_private->video_width = width;
-    g_mutex_unlock(&texture_rgba_private->mutex);
+    auto stride_align = fl_value_get_int(fl_value_lookup_string(args, "stride_align"));
+    auto texture_rgba = g_renderer_map[key];
+    FlutterRgbaRendererPluginOnRgba((void*)texture_rgba, data, data_length, width, height, stride_align);
     response = FL_METHOD_RESPONSE(fl_method_success_response_new(
         fl_value_new_bool(fl_texture_registrar_mark_texture_frame_available(self->texture_registrar, FL_TEXTURE(texture_rgba)))));
   }
@@ -96,7 +81,7 @@ static void texture_rgba_renderer_plugin_handle_method_call(
   {
     auto args = fl_method_call_get_args(method_call);
     auto key = fl_value_get_int(fl_value_lookup_string(args, "key"));
-    if (self->g_renderer_map->find(key) == self->g_renderer_map->end())
+    if (g_renderer_map.find(key) == g_renderer_map.end())
     {
       response = FL_METHOD_RESPONSE(fl_method_success_response_new(
           fl_value_new_int(0)));
@@ -104,7 +89,7 @@ static void texture_rgba_renderer_plugin_handle_method_call(
     else
     {
       // Return an address.
-      size_t rgba = reinterpret_cast<size_t>((void *)(*self->g_renderer_map)[key]);
+      size_t rgba = reinterpret_cast<size_t>((void *)g_renderer_map[key]);
       response = FL_METHOD_RESPONSE(fl_method_success_response_new(
           fl_value_new_int(rgba)));
     }
@@ -119,7 +104,6 @@ static void texture_rgba_renderer_plugin_handle_method_call(
 
 static void texture_rgba_renderer_plugin_dispose(GObject *object)
 {
-  TEXTURE_RGBA_RENDERER_PLUGIN(object)->g_renderer_map->clear();
   G_OBJECT_CLASS(texture_rgba_renderer_plugin_parent_class)->dispose(object);
 }
 
@@ -128,10 +112,7 @@ static void texture_rgba_renderer_plugin_class_init(TextureRgbaRendererPluginCla
   G_OBJECT_CLASS(klass)->dispose = texture_rgba_renderer_plugin_dispose;
 }
 
-static void texture_rgba_renderer_plugin_init(TextureRgbaRendererPlugin *self)
-{
-  self->g_renderer_map = new std::unordered_map<int64_t, TextureRgba *>();
-}
+static void texture_rgba_renderer_plugin_init(TextureRgbaRendererPlugin *self) {}
 
 static void method_call_cb(FlMethodChannel *channel, FlMethodCall *method_call,
                            gpointer user_data)
@@ -160,26 +141,25 @@ void texture_rgba_renderer_plugin_register_with_registrar(FlPluginRegistrar *reg
 
 
 extern "C" {
-   void FlutterRgbaRendererPluginOnRgba(void *texture_rgba_ptr, const uint8_t *buffer, int width, int height) {
-      TextureRgba* texture_rgba = TEXTURE_RGBA_RENDERER_TEXTURE_RGBA(texture_rgba_ptr);
-      auto priv = (TextureRgbaPrivate *)texture_rgba_get_instance_private(texture_rgba);
-      g_mutex_lock(&priv->mutex);
-      if (priv->texture_id != 0) {
+   void FlutterRgbaRendererPluginOnRgba(void *texture_rgba_ptr, const uint8_t *buffer, int len, int width, int height, int stride_align) {
+      TextureRgba* self = TEXTURE_RGBA(texture_rgba_ptr);
+      g_mutex_lock(&self->mutex);
+      // selfate has registered a texture_id,
+      if (self->texture_id != 0 && !g_atomic_int_get(&self->buffer_ready) && !texture_rgba_is_terminate(self)) {
         // copy data to the texture.
-        priv->buffer_length = 4 * width * height;
-        auto copied_data = new uint8_t[priv->buffer_length];
-        memcpy(copied_data, buffer, priv->buffer_length); 
-        if (priv->buffer != nullptr) {
-          delete[] priv->buffer;
-        }
-        priv->buffer = copied_data;
-        priv->video_height = height;
-        priv->video_width = width;
+        auto copied_data = new uint8_t[len];
+        memcpy(copied_data, buffer, len); 
+        switch_rgba(copied_data, len, height);
+        // It's safe to working on a non reading index
+        g_atomic_pointer_set(&self->buffer, copied_data);
+        g_atomic_int_set(&self->video_height, height);
+        g_atomic_int_set(&self->video_width, width);
+        g_atomic_int_set(&self->buffer_ready, TRUE);
         fl_texture_registrar_mark_texture_frame_available(
-          priv->texture_registrar,
-          FL_TEXTURE(texture_rgba)
+          self->texture_registrar,
+          FL_TEXTURE(self)
         );
       }
-      g_mutex_unlock(&priv->mutex);
+      g_mutex_unlock(&self->mutex);
    }
 };