+++ macos/Classes/TextureRgbaApi.m
@@ -13,10 +13,9 @@ extern "C" {
 #endif
     
 /// Keep the same name with Windows.
-void FlutterRgbaRendererPluginOnRgba(void* texture_rgba_ptr, const uint8_t* buffer, int width, int height) {
+void FlutterRgbaRendererPluginOnRgba(void* texture_rgba_ptr, const uint8_t* buffer, int len, int width, int height, int stride_align) {
     TextRgba* texture_rgba = (__bridge TextRgba *)(texture_rgba_ptr);
-    NSData* data = [NSData dataWithBytesNoCopy:(void*)buffer length:width * height * 4 freeWhenDone:FALSE];
-    [texture_rgba markFrameAvaliableWithData:data width:width height:height];
+    [texture_rgba markFrameAvaliableRawWithBuffer:buffer len:len width:width height:height stride_align: stride_align];
 }
 
 #if __cplusplus